web/requirements.txt
bottle==0.12.25
bottle-websocket==0.2.9
cffi==1.16.0
Eel==0.16.0
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
lxml==5.1.0
pycparser==2.21
pyparsing==3.1.2
python-docx==1.1.0
typing_extensions==4.10.0
whichcraft==0.6.1
zope.event==5.0
zope.interface==6.2
